rv64 system airs

+++ crates/vm/src/arch/testing/cuda.rs
@@ -206,8 +206,7 @@ impl TestBuilder<F> for GpuChipTestBuilder {
     ) -> (usize, usize) {
         let register = self.get_default_register(reg_increment);
         let pointer = self.get_default_pointer(pointer_increment);
-        // Cast to u32 to ensure we write exactly 4 bytes (RV32 register size).
-        self.write(1, register, (pointer as u32).to_le_bytes().map(F::from_u8));
+        self.write(1, register, (pointer as u64).to_le_bytes().map(F::from_u8));
         (register, pointer)
     }
 
@@ -337,11 +336,10 @@ impl GpuChipTestBuilder {
         pointer: usize,
         writes: Vec<[F; NUM_LIMBS]>,
     ) {
-        // Cast to u32 to ensure we write exactly 4 bytes (RV32 register size).
         self.write(
             1usize,
             register,
-            (pointer as u32).to_le_bytes().map(F::from_u8),
+            (pointer as u64).to_le_bytes().map(F::from_u8),
         );
         // Always write in DEFAULT_BLOCK_SIZE-byte chunks to match the fixed block size.
         for (i, &write) in writes.iter().enumerate() {
